Added Vertical Federated Goverment for Deep Learning
The vertical FL follows the structure of the Horizontal case: The FL is separated from Server and client nodes.
Still, we need to define how model's meta-parameters are accessed. Namely, the embeddings and gradients, don't really
define the model. Yet at the present stage, they are considered as (private) model's parameters.
Do we need to define multiple private queries for those instead?