FederatedData: Added a general method to loop on nodes
At initialization, class attributes are created dynamically from node's attributes names to loop on the federated data.
This way, the FederatedData object have all the (callable) node methods, but simply they are iterated over all nodes.
Note that this way they are hinted when you write "federated_data.+TAB", so the user is not required to type any string value and there is no way to make a mistake.